+++ src/client.js
@@ -12,6 +12,8 @@ const fs = require('fs');
 const FormData = require('form-data');
 const socket = require('socket.io-client');
 const { RequestError } = require("./errors");
+const bunyan = require("bunyan");
+const {CLIENT_SCHEMA} = require("./utils/evaluator");
 let axios;
 
 // Browser/Node axios import
@@ -55,70 +57,156 @@ class WideSkyClient {
     #clientId
     #clientSecret
     #accessToken
+    #logger
+    /**
+     * If this is true (default) then all http requests made by the client
+     * will have the 'Accept-Encoding' header with value of 'gzip, deflate'
+     * append to it.
+     */
+    _acceptGzipEncoding
+    _impersonate        // The user id which the original user is impersonating as.
 
-    constructor(base_uri,
+    /**
+     * Constructor for WideSky Client
+     * @param baseUri URI to access the WideSky API (excluding /api)
+     * @param username Username of the WideSky user to authenticate with.
+     * @param password Password of the WideSky user to authenticate with.
+     * @param clientId Client ID for OAuth 2.0 authentication.
+     * @param clientSecret Client secret for OAuth 2.0 authentication.
+     * @param logger A Bunyan logging instance.
+     * @param accessToken A valid WideSky access token.
+     * @param options A Object containing attributes "axios" and "client" for configuring the axios and WideSky client
+     *                respectively. Axios configurations are described at https://axios-http.com/docs/config_defaults.
+     */
+    constructor(baseUri,
                 username,
                 password,
                 clientId,
                 clientSecret,
-                log,
-                accessToken) {
-        this.base_uri = base_uri;
+                logger,
+                accessToken,
+                options={}) {
+        this.baseUri = baseUri;
         this.#username = username;
         this.#password = password;
         this.#clientId = clientId;
         this.#clientSecret = clientSecret;
-        this.#accessToken = accessToken
+        this.#accessToken = accessToken;
+        this.#logger = logger;
+        this.options = options;
+        this.clientOptions = null;
+        this._impersonate = null;
+        this._acceptGzipEncoding  = true;
+        this.initialised = false;
 
-        /* Logger instance */
-        this._log = log;
+        this.initAccessToken();
+        this.initAxios();
+        // Client option initiator is async. Wait for this to complete before submitting requests
+        this.initWaitFor = this.initClientOptions();
+    }
 
-        if (this.#accessToken) {
-            if (!this.#accessToken.hasOwnProperty('refresh_token') ||
-                !this.#accessToken.hasOwnProperty('expires_in') ||
-                !this.#accessToken.hasOwnProperty('token_type') ||
-                !this.#accessToken.hasOwnProperty('access_token')) {
-                throw new Error(`Parameter 'accessToken' is not a valid WideSky token.`);
+    /**
+     * Create a WideSky client from a given set of configurations
+     * @param config An Object defining the configurations for the WideSkyClient. Requires attributes serverURL,
+     *               username, password, clientId, clientSecret. Optional attributes are logger, accessToken and
+     *               options. If no logger is specified, a Bunyan logging instance will be created.
+     * @returns {WideSkyClient} A WideSky client instance.
+     */
+    static makeFromConfig(config={}) {
+        const requiredProps = [
+            "serverURL",
+            "username",
+            "password",
+            "clientId",
+            "clientSecret"
+        ];
+        for (const prop of requiredProps) {
+            if (config[prop] === undefined) {
+                throw new Error(`Configuration parameter requires properties ${requiredProps.join(", ")}`);
             }
         }
 
+        const {
+            serverURL,
+            username,
+            password,
+            clientId,
+            clientSecret,
+            accessToken,
+            options
+        } = config;
+        let { logger } = config;
+
+        if (logger === undefined) {
+            logger = bunyan.createLogger({
+                name: "WideSky-Client"
+            });
+        }
+
+        return new WideSkyClient(
+            serverURL,
+            username,
+            password,
+            clientId,
+            clientSecret,
+            logger,
+            accessToken,
+            options
+        );
+    }
+
+    /**
+     * Initialise the Client access token
+     */
+    initAccessToken() {
         /*
          * The authentication response, used for storing the access token and
          * refresh token.
          */
-        this._ws_token = this.#accessToken || null;
+        this._ws_token = null;
         /*
-         * Refresh token retrieval.  The list of waiters for a refresh token
-         * add themselves here.  If `null`, then no refresh is in progress.
+         * Refresh token retrieval. The list of waiters for a refresh token
+         * add themselves here. If `null`, then no refresh is in progress.
          */
         this._ws_token_wait = null;
 
-        /**
-         * The user id which the original user is impersonating as.
-         */
-        this._impersonate = null;
-
-        /**
-         * If this is true (default) then all http requests made by the client
-         * will have the 'Accept-Encoding' header with value of 'gzip, deflate'
-         * append to it.
-         *
-         * @type {boolean}
-         * @private
-         */
-        this._acceptGzipEncoding = true;
-
-        this.initAxios(base_uri);
+        if (this.#accessToken) {
+            for (const tokenProp of ['refresh_token', 'expires_in', 'token_type', 'access_token']) {
+                if (this.#accessToken[tokenProp] === undefined) {
+                    throw new Error(`Parameter 'accessToken' is not a valid WideSky token.`);
+                }
+            }
+            this._ws_token = this.#accessToken;
+        }
     }
 
     /**
      * Apply the config to be used for all axios requests.
-     * @param baseUri Uri of the target API server.
      */
-    initAxios(baseUri) {
-        this.axios = axios.create({
-            baseURL: baseUri
-        });
+    initAxios() {
+        this.axios = axios.create(Object.assign({
+            baseURL: this.baseUri
+        }, this.options.axios || {}));
+    }
+
+    /**
+     * Initialise the WideSkyClient with the user configurations.
+     * @returns {Promise<void>}
+     */
+    async initClientOptions() {
+        try {
+            await CLIENT_SCHEMA.validate(this.options.client);
+            this.clientOptions = CLIENT_SCHEMA.cast(this.options.client);
+            this.setAcceptGzip(this.clientOptions.acceptGzip);
+
+            if (this.clientOptions.impersonateAs !== null) {
+                this.impersonateAs(this.clientOptions.impersonateAs);
+            }
+
+            this.initialised = true;
+        } catch (error) {
+            throw new Error(error.message);
+        }
     }
 
     /**
@@ -130,11 +218,19 @@ class WideSkyClient {
     };
 
     impersonateAs(userId) {
-        this._impersonate = userId;
+        if (!this.initialised) {
+            const oldPromise = this.initWaitFor;
+            this.initWaitFor = new Promise(async (resolve) => {
+                await oldPromise;
+                this._impersonate = userId;
+            });
+        } else {
+            this._impersonate = userId;
+        }
     };
 
     isImpersonating() {
-        return !!this._impersonate;
+        return this._impersonate !== null;
     };
 
     unsetImpersonate() {
@@ -159,9 +255,14 @@ class WideSkyClient {
      * @returns Data from response of request.
      */
     async _wsRawSubmit(method, uri, body, config) {
+        if (!this.initialised) {
+            this.#logger.info("Not finished initialising. Waiting...");
+            await this.initWaitFor;
+        }
+
         /* istanbul ignore next */
-        if (this._log) {
-            this._log.trace(config, 'Raw request');
+        if (this.#logger) {
+            this.#logger.trace(config, 'Raw request');
         }
 
         let res;
@@ -248,8 +349,8 @@ class WideSkyClient {
      */
     _doLogin() {
         /* istanbul ignore next */
-        if (this._log) {
-            this._log.trace('Performing login attempt');
+        if (this.#logger) {
+            this.#logger.trace('Performing login attempt');
         }
 
         return this._wsRawSubmit(
@@ -275,7 +376,7 @@ class WideSkyClient {
      */
     _doRefresh() {
         /* istanbul ignore next */
-        if (this._log) this._log.trace('Performing token refresh attempt');
+        if (this.#logger) this.#logger.trace('Performing token refresh attempt');
 
         return this._wsRawSubmit(
             'POST',
@@ -295,8 +396,8 @@ class WideSkyClient {
 
     _getTokenSuccess(token, resolve) {
         /* istanbul ignore next */
-        if (this._log) {
-            this._log.info('Logged in to API server');
+        if (this.#logger) {
+            this.#logger.info('Logged in to API server');
         }
         this._ws_token = token;
 
@@ -311,8 +412,8 @@ class WideSkyClient {
 
     _getTokenFail(err, reject) {
         /* istanbul ignore next */
-        if (this._log) {
-            this._log.warn(err, 'Failed to log into API server');
+        if (this.#logger) {
+            this.#logger.warn(err, 'Failed to log into API server');
         }
         this._ws_token = null;
 
@@ -337,8 +438,8 @@ class WideSkyClient {
         if (this._ws_token_wait !== null) {
             /* Join the queue */
             /* istanbul ignore next */
-            if (this._log) {
-                this._log.trace('Waiting for token acquisition');
+            if (this.#logger) {
+                this.#logger.trace('Waiting for token acquisition');
             }
 
             return new Promise( (resolve, reject) => {
@@ -353,8 +454,8 @@ class WideSkyClient {
             /* No token, so acquire one */
             this._ws_token_wait = [];
             /* istanbul ignore next */
-            if (this._log) {
-                this._log.trace('Begin token acquisition');
+            if (this.#logger) {
+                this.#logger.trace('Begin token acquisition');
             }
 
             firstStep = this._doLogin();
@@ -362,8 +463,8 @@ class WideSkyClient {
         else if (this._ws_token.expires_in < Date.now()) {
             /* Token is expired, so do a refresh */
             /* istanbul ignore next */
-            if (this._log) {
-                this._log.trace('Begin token refresh');
+            if (this.#logger) {
+                this.#logger.trace('Begin token refresh');
             }
 
             this._ws_token_wait = [];
@@ -381,8 +482,8 @@ class WideSkyClient {
                     /* If we're refreshing, try a full log-in */
                     if (refresh) {
                         /* istanbul ignore next */
-                        if (this._log) {
-                            this._log.info(
+                        if (this.#logger) {
+                            this.#logger.info(
                                 err,
                                 'Refresh fails, trying log-in instead'
                             );
@@ -580,8 +681,8 @@ class WideSkyClient {
         /* Ensure updateRec lists `id` */
         if ((!present.id) && (op === 'updateRec')) {
             /* istanbul ignore next */
-            if (this._log) {
-                this._log.trace(entities, 'Entities lacks id column');
+            if (this.#logger) {
+                this.#logger.trace(entities, 'Entities lacks id column');
             }
 
             throw new Error('id is missing');
@@ -652,8 +753,8 @@ class WideSkyClient {
      */
     createUser(email, name, description, roles, password=null, method=AUTH_METHOD.LOCAL) {
         /* istanbul ignore next */
-        if (this._log) {
-            this._log.trace('Creating a new user: ' + email);
+        if (this.#logger) {
+            this.#logger.trace('Creating a new user: ' + email);
         }
 
         if (email.length < 1) {
@@ -689,8 +790,8 @@ class WideSkyClient {
      */
     updatePassword(newPassword) {
         /* istanbul ignore next */
-        if (this._log) {
-            this._log.trace('Updating password');
+        if (this.#logger) {
+            this.#logger.trace('Updating password');
         }
 
         if (!newPassword) {